fix: sharedPreferenceをcore/localフォルダ内に移動・コードのリファクタリングをしました
