+++ lib/ui/screen/detail/detail_post_view_model.dart
@@ -1,7 +1,7 @@
-import 'package:food_gram_app/core/config/shared_preference/shared_preference.dart';
 import 'package:food_gram_app/core/data/supabase/block_list.dart';
 import 'package:food_gram_app/core/data/supabase/post/delete_service.dart';
 import 'package:food_gram_app/core/data/supabase/post_stream.dart';
+import 'package:food_gram_app/core/local/shared_preference.dart';
 import 'package:food_gram_app/core/model/posts.dart';
 import 'package:food_gram_app/core/utils/provider/loading.dart';
 import 'package:food_gram_app/ui/screen/detail/detail_post_state.dart';
